% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createConclass.R
\name{createConclass}
\alias{createConclass}
\title{Creates conclass column using status and decay vectors.}
\usage{
createConclass(statusVec, decayVec)
}
\arguments{
\item{statusVec}{Vector of tree status values.}

\item{decayVec}{Vector of tree decay values.}
}
\description{
Creates conclass column using status and decay vectors.
}
