% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuleDecomp.R
\name{fuleDecomp}
\alias{fuleDecomp}
\title{Calculates Fule's decomposition rates}
\usage{
fuleDecomp(rate, percentiles)
}
\arguments{
\item{percentiles}{}
}
\description{
Calculates Fule's decomposition rates
}
