% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liveTreeAge.R
\name{liveTreeAge}
\alias{liveTreeAge}
\title{Predict unknown live tree ages from DBH of known live trees}
\usage{
liveTreeAge(
  data = data,
  speciesCol = "Species",
  ageCol = "Age",
  dbhCol = "DBH",
  statusCol = "Status",
  measYear = measYear
)
}
\arguments{
\item{data}{Data with species, age, DBH, and tree status columns.}

\item{speciesCol}{Species column name in data as a string.}

\item{ageCol}{Age column name in data as a string.}

\item{dbhCol}{DBH column name in data as a string.}

\item{statusCol}{Tree Status column name in data as a string.}

\item{measYear}{Year that data was measured.}
}
\value{
Data with predicted ages for all live trees.
}
\description{
Predict unknown live tree ages from DBH of known live trees
}
