% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standrecon.R
\name{standrecon}
\alias{standrecon}
\title{Reconstruct stand conditions from a given reference year}
\usage{
standrecon(
  data = data,
  measYear,
  refYear,
  avgIncVec,
  plotSize,
  nPlots = 1,
  barkEqList = list(),
  speciesCol = "Species",
  ageCol = "Age",
  dbhCol = "DBH",
  statusCol = "Status",
  decayCol = "Decay",
  percentiles = c(0.25, 0.5, 0.75),
  minDbh = 5
)
}
\arguments{
\item{data}{Data with species, age, DBH, tree status, and tree decay columns.}

\item{measYear}{Year that data was measured.}

\item{refYear}{Reference years to reconstruct stand conditions.}

\item{avgIncVec}{Vector of average increment growth per species with units mm/year. (PIEN = 0.5, ABBI = 0.3)}

\item{plotSize}{Size of plots in m².}

\item{nPlots}{Number of plots with plotSize.}

\item{barkEqList}{List of bark correction equations. Default bark corrections for species codes used otherwise. list(PIPO = function(x) x*1.1029 + 0.7162)}

\item{speciesCol}{Species column name in data as a string.}

\item{ageCol}{Species column name in data as a string.}

\item{dbhCol}{DBH column name in data as a string.}

\item{statusCol}{Species status column name in data as a string.}

\item{decayCol}{Species decay column name in data as a string.}

\item{percentiles}{Percentiles to be calculated for sensitivity analysis.}

\item{minDbh}{Minimum DBH threshold for stand calculations in cm.}
}
\description{
Reconstruct stand conditions from a given reference year
}
