% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liveTreeAge.R
\name{liveTreeAge}
\alias{liveTreeAge}
\title{Predict unknown live tree ages from DBH of known live trees}
\usage{
liveTreeAge(
  data,
  speciesCol = Species,
  ageCol = Age,
  dbhCol = DBH,
  statusCol = Status,
  measYear
)
}
\arguments{
\item{data}{Data with species, age, DBH, and tree status columns.}

\item{speciesCol}{Species column in data.}

\item{ageCol}{Age column in data.}

\item{dbhCol}{DBH column in data.}

\item{statusCol}{Tree Status column in data.}

\item{measYear}{Year that data was measured.}
}
\value{
Data with predicted ages for all live trees.
}
\description{
Predict unknown live tree ages from DBH of known live trees
}
