% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standrecon.R
\name{standrecon}
\alias{standrecon}
\title{Reconstruct stand conditions from a given reference year}
\usage{
standrecon(
  data,
  meas_year,
  ref_year,
  avg_inc_vec,
  plot_size,
  n_plots = 1,
  bark_eq_list = list(),
  species_col = "Species",
  age_col = "Age",
  dbh_col = "DBH",
  status_col = "Status",
  decay_col = "Decay",
  percentiles = c(0.25, 0.5, 0.75),
  min_dbh = 5
)
}
\arguments{
\item{data}{Data with species, age, DBH, tree status, and tree decay columns.}

\item{meas_year}{Year that data was measured.}

\item{ref_year}{Reference years to reconstruct stand conditions.}

\item{avg_inc_vec}{Vector of average increment growth per species with units mm/year. (PIEN = 0.5, ABBI = 0.3)}

\item{plot_size}{Size of plots in m².}

\item{n_plots}{Number of plots with plot_size.}

\item{bark_eq_list}{List of bark correction equations. Default bark corrections for species codes used otherwise. list(PIPO = function(x) x*1.1029 + 0.7162)}

\item{species_col}{Species column name in data as a string.}

\item{age_col}{Species column name in data as a string.}

\item{dbh_col}{DBH column name in data as a string.}

\item{status_col}{Species status column name in data as a string.}

\item{decay_col}{Species decay column name in data as a string.}

\item{percentiles}{Percentiles to be calculated for sensitivity analysis.}

\item{min_dbh}{Minimum DBH threshold for stand calculations in cm.}
}
\value{
A data frame with one row per species, reference year, and percentile. Columns include output type (reconstructed or measured),
reference year, percentile, species code, basal area per hectare, and stem density per hectare.
}
\description{
Reconstruct stand conditions from a given reference year
}
\examples{
data(standrecon_example_data)

out <- standrecon(
  data = standrecon_example_data,
  meas_year = 2025,
  ref_year = c(1950, 1975),
  avg_inc_vec = c(PIEN = 0.5, ABBI = 0.3, PIPO = 0.4),
  plot_size = 1000)

out
}
